More Changes

+++ StudioUp.Repo/Repositories/CustomerRepository.cs
@@ -44,6 +44,10 @@ namespace StudioUp.Repo.Repositories
             try
             {
                 var c = await context.Customers.FirstOrDefaultAsync(t => t.Id == id);
+                if (c == null)
+                {
+                    return false;
+                }
                 var mapC = mapper.Map<Customer>(c);
                 context.Customers.Remove(mapC);
                 await context.SaveChangesAsync();
@@ -51,7 +55,8 @@ namespace StudioUp.Repo.Repositories
             }
             catch (Exception ex)
             {
-                return false;
+                throw new Exception("Cann't delete this object");
+
             }
         }
 
@@ -77,6 +82,7 @@ namespace StudioUp.Repo.Repositories
             try
             {
                 var c = await context.Customers.FirstOrDefaultAsync(t => t.Id == id);
+
                 var mapCust = mapper.Map<CustomerDTO>(c);
                 return mapCust;
 
@@ -104,7 +110,7 @@ namespace StudioUp.Repo.Repositories
                 customerToUpdate.PaymentOptionId = entity.PaymentOptionId;
                 customerToUpdate.HMOId = entity.HMOId;
                 customerToUpdate.CustomerTypeId = entity.CustomerTypeId;
-                //customerToUpdate.IsActive = entity.IsActive;
+                customerToUpdate.IsActive = entity.IsActive;
                 customerToUpdate.SubscriptionTypeId = entity.SubscriptionTypeId;
                 customerToUpdate.Tel = entity.Tel;
                 context.Customers.Update(mapper.Map<Customer>(customerToUpdate));
@@ -115,7 +121,7 @@ namespace StudioUp.Repo.Repositories
             }
             catch (Exception ex)
             {
-                throw new Exception("Not secceed");
+                throw new Exception("Cann't update this object");
             }
         }
     }
